+++ obj/Debug/net8.0/BackendServer.AssemblyInfo.cs
@@ -13,7 +13,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("BackendServer")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+b1a53c3d59af6d4775bf94c595283abaaa0dce09")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+e9edec13e34ee5db48556bfe0fb69042af0bc3e2")]
 [assembly: System.Reflection.AssemblyProductAttribute("BackendServer")]
 [assembly: System.Reflection.AssemblyTitleAttribute("BackendServer")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]