+++ CHANGELOG.md
@@ -32,5 +32,8 @@ All notable changes to this project will be documented in this file.
 
 ## [1.2.1] - 2026-02-18
 
-### Added
+### Fixed
 -  Prevent unbounded dictionary growth in rate limit middleware
+
+### Added
+-  Health check endpoint